feat: add `Position` type
